Barrier Interface (Basic)
Created a basic implementation of the barrier user interface using
Swing.

Implemented button which enter a registration using a JTextField;
Implemented a button to allow the vehicle to pass (Simulated);
Implemented JTextField to enter a registration with use of the
actionListener interface.
Added the ActionListener interface to the class to allow listening for
buttons;