Refactor Pass #1
Timer
Changed if statement in line 117 to use .equals instead of == to be in
line with everyone else using .equals;
Refactored constructor code to use a new method to make the UI elements.
- loadGUI();

Campus_security
Refactored constructor code to use a new method to make the UI elements.
- loadGUI();

Barrier
Refactored constructor code to use a new method to make the UI elements.
- loadGUI();

Administration_office
Refactored constructor code to use a new method to make the UI elements.
- loadGUI();
Noticed code that I am not sure why it is there, line 95 - permitList =
new Permit_list(); Is this needed as we pass in a permitList in
constructor?

ChoicePane
Made a method called setDialogText(String, String, String) to avoid the
repeating code in the switch statement - line 38;

Overall
Changed the location of the panels to line up better with each other,
allowing space for extra barriers should they be needed.