Changes to Timer, and Permits.
-Created a getPermitsByType method which returns an ArrayList of type
Permit. Takes a parameter of a String which will be used as a permit
type.

Using the ForEach method in HashTable I am iterating through it,
comparing the class.toString against the permit type. Adding it to an
ArrayList if it is of that type.

-Created an abstract getDate method in Permit which is then inherited by
all sub classes to collect the date from them if needed.

-Created cancelOutOfDatePermits which checks the date of the returned
ArrayList against the current date. If they have expried they are
deleted.

-TODO
Add any other permits which expire
Fix the combo boxes to display the permits
Update the combo boxes after a permit has expired.