Validation: Registration #30
Fixed a better validation for registration plates in the validate
method.

Done some general tidying up of code and got some stuff in place for the
#26 issue

Fixed an issue that the permits were not updating after creating a new
permit. This was caused by a missed break; in the switch statement in
the createPermit method.